+++ packages/abap-deploy-config-inquirer/src/prompts/questions/config/package.ts
@@ -6,7 +6,7 @@ import {
 } from '../../conditions';
 import { t } from '../../../i18n';
 import { getSystemConfig } from '../../../utils';
-import { getPackageChoices, getPackageInputChoices, shouldValidatePackage } from '../../helpers';
+import { getPackageChoices, getPackageInputChoices, shouldRunValidation } from '../../helpers';
 import { defaultPackage, defaultPackageChoice } from '../../defaults';
 import { validatePackageChoiceInput, validatePackageChoiceInputForCli, validatePackage } from '../../validators';
 import {
@@ -93,13 +93,15 @@ export function getPackagePrompts(
             default: (previousAnswers: AbapDeployConfigAnswersInternal): string =>
                 defaultPackage(previousAnswers.packageManual || options.packageManual?.default, options?.packageManual),
             validate: async (input: string, answers: AbapDeployConfigAnswersInternal): Promise<boolean | string> => {
-                if (shouldValidatePackage(prevAnswers as AbapDeployConfigAnswersInternal, answers)) {
+                if (shouldRunValidation(prevAnswers as AbapDeployConfigAnswersInternal, answers)) {
                     prevValidationResult = await validatePackage(
                         input,
                         answers,
                         options.packageManual,
                         options.ui5AbapRepo,
-                        options.backendTarget
+                        options.backendTarget,
+                        undefined,
+                        options.adpProjectType
                     );
                 }
                 prevAnswers = answers;
@@ -146,13 +148,15 @@ export function getPackagePrompts(
                 const pkgValue: string = (input as ListChoiceOptions)?.value
                     ? (input as ListChoiceOptions).value
                     : input;
-                if (shouldValidatePackage(prevAnswers as AbapDeployConfigAnswersInternal, answers)) {
+                if (shouldRunValidation(prevAnswers as AbapDeployConfigAnswersInternal, answers)) {
                     prevValidationResult = await validatePackage(
                         pkgValue,
                         answers,
                         options.packageAutocomplete,
                         options.ui5AbapRepo,
-                        options.backendTarget
+                        options.backendTarget,
+                        undefined,
+                        options.adpProjectType
                     );
                 }
                 prevAnswers = answers;