+++ packages/adp-tooling/src/writer/project-utils.ts
@@ -17,7 +17,7 @@ import {
     enhanceUI5YamlWithCustomTask,
     enhanceUI5YamlWithTranspileMiddleware,
     enhanceUI5YamlWithCfCustomTask,
-    enhanceUI5YamlWithCfCustomMiddleware
+    enhanceUI5YamlWithFioriToolsMiddleware
 } from './options';
 
 import type { Package } from '@sap-ux/project-access';
@@ -182,8 +182,10 @@ export async function writeCfUI5Yaml(projectPath: string, data: CfAdpWriterConfi
         enhanceUI5YamlWithCustomConfig(ui5Config, data.customConfig);
         /** Builder task */
         enhanceUI5YamlWithCfCustomTask(ui5Config, data);
+
         /** Middlewares */
-        enhanceUI5YamlWithCfCustomMiddleware(ui5Config, data);
+        // Add fiori-tools-proxy and fiori-tools-preview for local development
+        enhanceUI5YamlWithFioriToolsMiddleware(ui5Config);
 
         fs.write(ui5ConfigPath, ui5Config.toString());
     } catch (e) {