+++ packages/cf-deploy-config-writer/src/index.ts
@@ -1,4 +1,13 @@
 export * from './mta-config';
 export * from './cf-writer';
-export { DefaultMTADestination } from './constants';
+export {
+    DefaultMTADestination,
+    MAX_MTA_ID_LENGTH,
+    MAX_MTA_PREFIX_LENGTH,
+    MAX_MTA_PREFIX_SHORT_LENGTH,
+    MAX_MTA_PREFIX_SHORTER_LENGTH,
+    MAX_ABAP_SERVICE_PREFIX_LENGTH,
+    MAX_ABAP_SERVICE_NAME_LENGTH,
+    MTA_FILE_OPERATION_DELAY_MS
+} from './constants';
 export { CFBaseConfig, CFAppConfig, CAPConfig, RouterModuleType, ApiHubConfig, ApiHubType } from './types';