+++ packages/btp-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/btp-utils
 
+## 1.1.9
+
+### Patch Changes
+
+- dd2131c: Axios upgrade from bas-sdk
+
+## 1.1.8
+
+### Patch Changes
+
+- 2fc459c: Upgrade axios
+
 ## 1.1.7
 
 ### Patch Changes