+++ packages/control-property-editor-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux-private/control-property-editor-common
 
+## 0.7.6
+
+### Patch Changes
+
+- cc1c422: fix(deps): update dependency npm-run-all2 to v8
+
+## 0.7.5
+
+### Patch Changes
+
+- d588c26: fix(deps): update dependency rimraf to v6.1.3
+
 ## 0.7.4
 
 ### Patch Changes