+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,117 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.9.22
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.9
+- @sap-ux/cds-odata-annotation-converter@0.7.8
+
+## 0.9.21
+
+### Patch Changes
+
+- cc1c422: fix(deps): update dependency npm-run-all2 to v8
+- Updated dependencies [0ecc5f1]
+- Updated dependencies [cc1c422]
+    - @sap-ux/project-access@1.35.8
+    - @sap-ux/cds-annotation-parser@0.2.31
+    - @sap-ux/cds-odata-annotation-converter@0.7.8
+    - @sap-ux/odata-annotation-core-types@0.5.5
+    - @sap-ux/odata-annotation-core@0.2.15
+    - @sap-ux/odata-entity-model@0.3.6
+    - @sap-ux/odata-vocabularies@0.4.26
+    - @sap-ux/xml-odata-annotation-converter@0.4.9
+
+## 0.9.20
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.7
+- @sap-ux/cds-odata-annotation-converter@0.7.7
+
+## 0.9.19
+
+### Patch Changes
+
+- Updated dependencies [cbd340a]
+    - @sap-ux/cds-odata-annotation-converter@0.7.7
+
+## 0.9.18
+
+### Patch Changes
+
+- Updated dependencies [bb310dc]
+- Updated dependencies [e5bc3ca]
+    - @sap-ux/project-access@1.35.6
+    - @sap-ux/cds-odata-annotation-converter@0.7.6
+
+## 0.9.17
+
+### Patch Changes
+
+- Updated dependencies [d57cc47]
+    - @sap-ux/project-access@1.35.5
+    - @sap-ux/cds-odata-annotation-converter@0.7.6
+
+## 0.9.16
+
+### Patch Changes
+
+- Updated dependencies [1fa3bb7]
+    - @sap-ux/odata-vocabularies@0.4.25
+    - @sap-ux/cds-annotation-parser@0.2.30
+    - @sap-ux/cds-odata-annotation-converter@0.7.6
+    - @sap-ux/xml-odata-annotation-converter@0.4.8
+
+## 0.9.15
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.4
+- @sap-ux/cds-odata-annotation-converter@0.7.5
+
+## 0.9.14
+
+### Patch Changes
+
+- Updated dependencies [346f09c]
+    - @sap-ux/project-access@1.35.3
+    - @sap-ux/cds-odata-annotation-converter@0.7.5
+
+## 0.9.13
+
+### Patch Changes
+
+- Updated dependencies [d2b772d]
+    - @sap-ux/project-access@1.35.2
+    - @sap-ux/cds-odata-annotation-converter@0.7.5
+
+## 0.9.12
+
+### Patch Changes
+
+- Updated dependencies [2fc459c]
+    - @sap-ux/odata-vocabularies@0.4.24
+    - @sap-ux/cds-annotation-parser@0.2.29
+    - @sap-ux/cds-odata-annotation-converter@0.7.5
+    - @sap-ux/xml-odata-annotation-converter@0.4.8
+    - @sap-ux/project-access@1.35.1
+
+## 0.9.11
+
+### Patch Changes
+
+- 3795bb2: Add Node.js engine requirement (>=20.x)
+- Updated dependencies [3795bb2]
+    - @sap-ux/cds-odata-annotation-converter@0.7.4
+    - @sap-ux/xml-odata-annotation-converter@0.4.8
+    - @sap-ux/odata-annotation-core-types@0.5.4
+    - @sap-ux/cds-annotation-parser@0.2.28
+    - @sap-ux/odata-annotation-core@0.2.14
+    - @sap-ux/odata-entity-model@0.3.5
+    - @sap-ux/odata-vocabularies@0.4.23
+
 ## 0.9.10
 
 ### Patch Changes