+++ packages/fiori-docs-embeddings/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fiori-docs-embeddings
 
+## 0.4.15
+
+### Patch Changes
+
+- d57cc47: Update fast-xml-parser
+
+## 0.4.14
+
+### Patch Changes
+
+- 3795bb2: Add Node.js engine requirement (>=20.x)
+
 ## 0.4.13
 
 ### Patch Changes