+++ packages/jest-runner-puppeteer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/jest-runner-puppeteer
 
+## 0.2.7
+
+### Patch Changes
+
+- 469c90c: fix(deps): update dependency which to v6
+
+## 0.2.6
+
+### Patch Changes
+
+- d588c26: fix(deps): update dependency rimraf to v6.1.3
+
+## 0.2.5
+
+### Patch Changes
+
+- 3795bb2: Add Node.js engine requirement (>=20.x)
+
 ## 0.2.4
 
 ### Patch Changes