+++ packages/nodejs-utils/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/nodejs-utils
 
+## 0.2.16
+
+### Patch Changes
+
+- bb310dc: fix(deps): update dependency semver to v7.7.4
+
+## 0.2.15
+
+### Patch Changes
+
+- Updated dependencies [dd2131c]
+    - @sap-ux/btp-utils@1.1.9
+
+## 0.2.14
+
+### Patch Changes
+
+- Updated dependencies [2fc459c]
+    - @sap-ux/btp-utils@1.1.8
+
 ## 0.2.13
 
 ### Patch Changes